requirements.txt
torch
transformers
opencv-python
matplotlib
numpy
tqdm
Pillow
